+++ packages/core/bootstrap/src/index.ts
@@ -153,6 +153,9 @@ const withMetrics: Middleware = async (execute, context) => async (input: Adapte
     record({
       statusCode: providerStatusCode ? 200 : 500,
       providerStatusCode,
+      type: providerStatusCode
+        ? metrics.HttpRequestType.DATA_PROVIDER_HIT
+        : metrics.HttpRequestType.ADAPTER_ERROR,
     })
     throw error
   }