+++ packages/sources/chain-reserve-wallet/src/endpoint/wallet.ts
@@ -2,14 +2,14 @@ import { Requester, Validator } from '@chainlink/ea-bootstrap'
 import { AdapterResponse, Config, ExecuteWithConfig, InputParameters } from '@chainlink/types'
 import { ethers } from 'ethers'
 
-const networks = ['cardano', 'doge']
+const networks = ['cardano', 'dogecoin']
 const chainIds = ['mainnet', 'testnet']
 
 const networkChainMap: { [key: string]: number } = {
   cardano_testnet: 0,
-  doge_testnet: 1,
+  dogecoin_testnet: 1,
   cardano_mainnet: 2,
-  doge_mainnet: 3,
+  dogecoin_mainnet: 3,
 }
 
 const isNetwork = (maybeNetwork: string) => networks.indexOf(maybeNetwork) !== -1